class-32/demo/blog-api-permissions/pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = ""
authors = ["Ahmad Alawad <ahmad.alawad.sf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
numpy = "^1.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
